Added Dockerfile, docker_readme, and docker_main
This commit introduces three new files to the project: Dockerfile, docker_readme.html, and docker_main.sh. These files are essential for setting up and running the project within a Docker container.

1) Dockerfile
The Dockerfile is a text document that contains a set of instructions for building a Docker image. It specifies the base image, sets up the necessary dependencies, and defines the commands to be executed when the container is run.

2) docker_readme
The docker_readme file is an HTML document that serves as the documentation page for the Docker setup instructions. It provides step-by-step guidance on how to set up the necessary environment and run the project using Docker.

3) docker_main
The docker_main file contains a Bash script that handles device connection and runs the Flutter app within the Docker container. It checks for device connectivity, establishes a connection with the specified IP address, and launches the Flutter app using the provided commands.